move question to separate js file and connect to index, initialize, still need licenses
