+++ utils/generateMarkdown.js
@@ -12,7 +12,8 @@ function renderLicenseSection(license) {}
 
 // TODO: Create a function to generate markdown for README
 function generateMarkdown(data) {
-  return `# ${data.title}
+  return `# ${data.fileName}
+
   ## Description
   ${data.description}
